refactor(onchain-interactions/blocksense-contracts): Improve `CLAggregatorAdapter`
- Make `getViemChain` reusable by exporting it to `common.ts`
